chore: storybook-config 설정에 맞춰 apps/frontend storybook 설정 수정
