+++ apps/frontend/.storybook/preview.tsx
@@ -1,0 +1,7 @@
+import basicPreview from "@repo/storybook-config/preview";
+
+const preview = {
+  ...basicPreview,
+};
+
+export default preview;
